# Faster Go version