I ported the C version to Go. Notably, I switched from Go's stdlib random number generator to the xorshift version from C. I also used 64-bit floats instead of 32-bit. I'm not actually sure how to verify that it is correct, but I think it was an interesting exercise and (to the extent it is correct) could make the benchmark more useful.

I reran the benchmarks I could with later implementations (on my 2015 MBP):

```
C        (gcc -O3):       23.8s
Julia    (julia 1.1.0):   32.8s
Go (alt) (go 1.12):       39.3s
Java     (java 1.8.0_60): 44.2s
Go (org) (go 1.12):       64.8s
OCaml    (ocaml 4.07.1):  79.1s
JS       (node 11.14.0):  137.0s
Pypy     (pypy 6.0.0):    139.4s
C#       (mono 4.2.1):    187.3s
```