@niofis 2019-04-26
thank you very much for this much-needed change, I was very disappointed with how fast go was before, this is a nice change